update proj structure
