+++ firmware/bootloader/SW4STM32/mchf_bootloader/.cproject
@@ -51,6 +51,7 @@
 									<listOptionValue builtIn="false" value="../../../proc/ui"/>
 									<listOptionValue builtIn="false" value="../../../proc/ui/lcd_driver"/>
 									<listOptionValue builtIn="false" value="../../../proc/ipc"/>
+									<listOptionValue builtIn="false" value="../../../proc/selftest"/>
 									<listOptionValue builtIn="false" value="../../../../"/>
 									<listOptionValue builtIn="false" value="../../../../STM32CubeH7/Drivers/CMSIS/Device/ST/STM32H7xx/Include"/>
 									<listOptionValue builtIn="false" value="../../../../STM32CubeH7/Drivers/STM32H7xx_HAL_Driver/Inc"/>
