+++ firmware/bootloader/SW4STM32/mchf_bootloader/version.h
@@ -23,6 +23,6 @@
 #define	MCHF_L_VER_MAJOR			0
 #define	MCHF_L_VER_MINOR			0
 #define	MCHF_L_VER_RELEASE			1
-#define	MCHF_L_VER_BUILD			48
+#define	MCHF_L_VER_BUILD			49
 // -----------------------------------------------------------------------------
 #endif
