+++ firmware/bootloader/proc/ui/ui_proc.h
@@ -1,0 +1,6 @@
+#ifndef __UI_PROC_H
+#define __UI_PROC_H
+
+void boot_process(void);
+
+#endif
